# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Integration Runtime",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/credentials/{provider_id}": {
      "put": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Set Credential",
        "operationId": "setCredential",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._-]{0,99}$",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_SetCredential_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credentials/{provider_id}/delete": {
      "post": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Delete Credential",
        "operationId": "deleteCredential",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._-]{0,99}$",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_DeleteCredential_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/storage": {
      "put": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Storage",
        "operationId": "configureStorage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_ConfigureStorage_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/stage-models": {
      "get": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Stages",
        "operationId": "getStageModels",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_StageModel__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Stage",
        "operationId": "configureStageModel",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_ConfigureStage_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/connection-checks": {
      "post": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Check",
        "operationId": "checkConnection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_ConnectionCheck_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "T03 settings"
        ],
        "summary": "Job",
        "operationId": "getGlobalJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Job_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeySecret": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "api_key",
            "title": "Kind"
          },
          "apiKey": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 1,
            "title": "Apikey",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "apiKey"
        ],
        "title": "ApiKeySecret"
      },
...
      "Command_ConfigureStage_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/ConfigureStage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[ConfigureStage]"
      },
      "Command_ConfigureStorage_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/ConfigureStorage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[ConfigureStorage]"
      },
...
      "Command_ConnectionCheck_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/ConnectionCheck"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[ConnectionCheck]"
      },
...
      "Command_DeleteCredential_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/DeleteCredential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[DeleteCredential]"
      },
...
      "Command_SetCredential_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/SetCredential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[SetCredential]"
      },
...
      "ConfigureStage": {
        "properties": {
          "phase": {
            "$ref": "#/components/schemas/PipelinePhase"
          },
          "capabilityId": {
            "$ref": "#/components/schemas/Uuid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "phase",
          "capabilityId"
        ],
        "title": "ConfigureStage"
      },
      "ConfigureStorage": {
        "properties": {
          "providerId": {
            "$ref": "#/components/schemas/ProviderId"
          },
          "region": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Region"
          },
          "bucket": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Bucket"
          },
          "credentialRef": {
            "$ref": "#/components/schemas/Uuid"
          },
          "retentionHours": {
            "$ref": "#/components/schemas/SafeInteger",
            "ge": 1,
            "le": 168
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "providerId",
          "region",
          "bucket",
          "credentialRef",
          "retentionHours"
        ],
        "title": "ConfigureStorage"
      },
...
      "ConnectionCheck": {
        "properties": {
          "capabilityId": {
            "$ref": "#/components/schemas/Uuid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capabilityId"
        ],
        "title": "ConnectionCheck"
      },
...
      "DeleteCredential": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirmed"
        ],
        "title": "DeleteCredential"
      },
...
      "Envelope_Job_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Job]"
      },
...
      "Envelope_MutationReceipt_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/MutationReceipt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[MutationReceipt]"
      },
...
      "Envelope_list_StageModel__": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/StageModel"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[list[StageModel]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "import",
              "probe",
              "animatic",
              "export",
              "diagnostic",
              "local_check",
              "connection_check"
            ],
            "title": "Kind"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "resultId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "state",
          "progress",
          "resultId",
          "errorCode"
        ],
        "title": "Job"
      },
...
      "MutationReceipt": {
        "properties": {
          "operationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "committedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "resourceId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "state": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operationId",
          "committedRevision",
          "resourceId",
          "state"
        ],
        "title": "MutationReceipt"
      },
...
      "OssSecret": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "oss",
            "title": "Kind"
          },
          "accessKeyId": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Accesskeyid",
            "writeOnly": true
          },
          "accessKeySecret": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 1,
            "title": "Accesskeysecret",
            "writeOnly": true
          },
          "securityToken": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16384,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Securitytoken",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "accessKeyId",
          "accessKeySecret",
          "securityToken"
        ],
        "title": "OssSecret"
      },
      "Persistence": {
        "type": "string",
        "enum": [
          "dpapi",
          "session_only"
        ]
      },
      "PipelinePhase": {
        "type": "string",
        "enum": [
          "story_adaptation",
          "story_outline",
          "story_scene",
          "story_dialogue",
          "image_character",
          "image_location",
          "image_prop",
          "image_keyframe",
          "video",
          "speech",
          "lipsync",
          "music",
          "sfx",
          "check"
        ]
      },
...
      "ProviderId": {
        "type": "string",
        "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._-]{0,99}$"
      },
...
      "SafeInteger": {
        "type": "integer",
        "ge": 0,
        "le": 9007199254740991
      },
...
      "SetCredential": {
        "properties": {
          "secret": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApiKeySecret"
              },
              {
                "$ref": "#/components/schemas/OssSecret"
              }
            ],
            "title": "Secret",
            "writeOnly": true,
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "api_key": "#/components/schemas/ApiKeySecret",
                "oss": "#/components/schemas/OssSecret"
              }
            }
          },
          "persistence": {
            "$ref": "#/components/schemas/Persistence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "secret",
          "persistence"
        ],
        "title": "SetCredential"
      },
...
      "StageModel": {
        "properties": {
          "phase": {
            "$ref": "#/components/schemas/PipelinePhase"
          },
          "capabilityId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "capabilityVersion": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Capabilityversion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "phase",
          "capabilityId",
          "capabilityVersion"
        ],
        "title": "StageModel"
      },
...
      "Uuid": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "format": "uuid"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "RuntimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "32-byte random token, base64url without padding"
      }
    }
  },
  "security": [
    {
      "RuntimeBearer": []
    }
  ]
}